база перед 7 заданием

+++ app/src/main/java/com/nickolay/android2ver2/database/CityData.kt
@@ -14,5 +14,5 @@ data class CityData(
         @ColumnInfo(name = "city_tempricha")
         var cityTempricha: Float,
         @ColumnInfo(name = "city_data")
-        var cityData: Int
+        var cityData: Long
     )
